Move workloadupdatestrategy to outer scope
