+++ tests/install_upgrade_operators/launcher_updates/constants.py
@@ -1,6 +1,6 @@
 import copy
 
-from tests.install_upgrade_operators.constants import WORKLOADUPDATEMETHODS
+from tests.install_upgrade_operators.constants import WORKLOAD_UPDATE_STRATEGY_KEY_NAME, WORKLOADUPDATEMETHODS
 from utilities.constants import LIVE_MIGRATE
 
 DEFAULT_BATCH_EVICTION_INTERVAL = "1m0s"
@@ -44,6 +44,5 @@ MOD_CUST_DEFAULT_BATCH_EVICTION_INTERVAL["batchEvictionInterval"] = DEFAULT_BATC
 MOD_CUST_DEFAULT_BATCH_EVICTION_SIZE["batchEvictionSize"] = DEFAULT_BATCH_EVICTION_SIZE
 MOD_CUST_DEFAULT_WORKLOAD_UPDATE_METHOD[WORKLOADUPDATEMETHODS] = DEFAULT_WORKLOAD_UPDATE_METHODS
 
-WORKLOAD_UPDATE_STRATEGY_KEY_NAME = "workloadUpdateStrategy"
 CUSTOM_STRATEGY = {WORKLOAD_UPDATE_STRATEGY_KEY_NAME: CUSTOM_WORKLOAD_UPDATE_STRATEGY}
 CUSTOM_WORKLOAD_STRATEGY_SPEC = {"spec": CUSTOM_STRATEGY}
