+++ tests/install_upgrade_operators/launcher_updates/test_reset_custom_values.py
@@ -1,12 +1,11 @@
 import pytest
 
-from tests.install_upgrade_operators.constants import WORKLOADUPDATEMETHODS
+from tests.install_upgrade_operators.constants import WORKLOAD_UPDATE_STRATEGY_KEY_NAME, WORKLOADUPDATEMETHODS
 from tests.install_upgrade_operators.launcher_updates.constants import (
     DEFAULT_WORKLOAD_UPDATE_STRATEGY,
     MOD_CUST_DEFAULT_BATCH_EVICTION_INTERVAL,
     MOD_CUST_DEFAULT_BATCH_EVICTION_SIZE,
     MOD_CUST_DEFAULT_WORKLOAD_UPDATE_METHOD,
-    WORKLOAD_UPDATE_STRATEGY_KEY_NAME,
 )
 from tests.install_upgrade_operators.utils import wait_for_spec_change
 from utilities.hco import get_hco_spec
