+++ tests/install_upgrade_operators/launcher_updates/test_update_default_workload_strategy.py
@@ -1,6 +1,6 @@
 import pytest
 
-from tests.install_upgrade_operators.constants import WORKLOADUPDATEMETHODS
+from tests.install_upgrade_operators.constants import WORKLOAD_UPDATE_STRATEGY_KEY_NAME, WORKLOADUPDATEMETHODS
 from tests.install_upgrade_operators.launcher_updates.constants import (
     DEFAULT_BATCH_EVICTION_INTERVAL,
     DEFAULT_BATCH_EVICTION_SIZE,
@@ -9,7 +9,6 @@ from tests.install_upgrade_operators.launcher_updates.constants import (
     MOD_CUST_DEFAULT_BATCH_EVICTION_INTERVAL,
     MOD_CUST_DEFAULT_BATCH_EVICTION_SIZE,
     MOD_CUST_DEFAULT_WORKLOAD_UPDATE_METHOD,
-    WORKLOAD_UPDATE_STRATEGY_KEY_NAME,
 )
 from tests.install_upgrade_operators.utils import wait_for_spec_change
 from utilities.hco import get_hco_spec
